tests: refresh generated references

+++ tests/golden/op_conv_conv.c
@@ -86,7 +86,7 @@ static inline void node0_conv(const float input0[1][1][4][4], const float weight
                 const idx_t oc_global = g * 1 + oc;
                 for (idx_t od0 = 0; od0 < 4; ++od0) {
                     for (idx_t od1 = 0; od1 < 4; ++od1) {
-                        float acc = bias[oc_global];
+                        double acc = (double)bias[oc_global];
                         for (idx_t ic = 0; ic < 1; ++ic) {
                             const idx_t ic_global = g * 1 + ic;
                             for (idx_t kd0 = 0; kd0 < 3; ++kd0) {
@@ -94,12 +94,12 @@ static inline void node0_conv(const float input0[1][1][4][4], const float weight
                                     const idx_t id0 = od0 * 1 + kd0 * 1 - 1;
                                     const idx_t id1 = od1 * 1 + kd1 * 1 - 1;
                                     if (id0 >= 0 && id0 < 4 && id1 >= 0 && id1 < 4) {
-                                        acc += input0[n][ic_global][id0][id1] * weights[oc_global][ic][kd0][kd1];
+                                        acc += (double)input0[n][ic_global][id0][id1] * (double)weights[oc_global][ic][kd0][kd1];
                                     }
                                 }
                             }
                         }
-                        output[n][oc_global][od0][od1] = acc;
+                        output[n][oc_global][od0][od1] = (float)acc;
                     }
                 }
             }
